+++ crates/client/rpc/src/methods/read/get_nonce.rs
@@ -23,6 +23,13 @@ use crate::Starknet;
 /// specific issue.
 
 pub fn get_nonce(starknet: &Starknet, block_id: BlockId, contract_address: Felt) -> StarknetRpcResult<Felt> {
+    // Check if block exists. We have to return a different error in that case.
+    let block_exists =
+        starknet.backend.contains_block(&block_id).or_internal_server_error("Checking if block is in database")?;
+    if !block_exists {
+        return Err(StarknetRpcApiError::BlockNotFound);
+    }
+
     if !starknet
         .backend
         .is_contract_deployed_at(&block_id, &contract_address)
@@ -43,15 +50,13 @@ pub fn get_nonce(starknet: &Starknet, block_id: BlockId, contract_address: Felt)
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::test_utils::{make_sample_chain_2, open_testing, SampleChain2};
+    use crate::test_utils::{sample_chain_for_state_updates, SampleChainForStateUpdates};
     use rstest::rstest;
     use starknet_core::types::BlockTag;
 
     #[rstest]
-    fn test_get_nonce() {
-        let _ = env_logger::builder().is_test(true).try_init();
-        let (backend, rpc) = open_testing();
-        let SampleChain2 { contracts, .. } = make_sample_chain_2(&backend);
+    fn test_get_nonce(sample_chain_for_state_updates: (SampleChainForStateUpdates, Starknet)) {
+        let (SampleChainForStateUpdates { contracts, .. }, rpc) = sample_chain_for_state_updates;
 
         // Block 0
         let block_n = BlockId::Number(0);
@@ -76,5 +81,18 @@ mod tests {
         assert_eq!(get_nonce(&rpc, block_n, contracts[0]).unwrap(), 3.into());
         assert_eq!(get_nonce(&rpc, block_n, contracts[1]).unwrap(), 2.into());
         assert_eq!(get_nonce(&rpc, block_n, contracts[2]).unwrap(), 2.into());
+    }
+
+    #[rstest]
+    fn test_get_nonce_not_found(sample_chain_for_state_updates: (SampleChainForStateUpdates, Starknet)) {
+        let (SampleChainForStateUpdates { contracts, .. }, rpc) = sample_chain_for_state_updates;
+
+        // Not found
+        let block_n = BlockId::Number(3);
+        assert_eq!(get_nonce(&rpc, block_n, contracts[0]), Err(StarknetRpcApiError::BlockNotFound));
+        let block_n = BlockId::Number(0);
+        assert_eq!(get_nonce(&rpc, block_n, contracts[1]), Err(StarknetRpcApiError::ContractNotFound));
+        let does_not_exist = Felt::from_hex_unchecked("0x7128638126378");
+        assert_eq!(get_nonce(&rpc, block_n, does_not_exist), Err(StarknetRpcApiError::ContractNotFound));
     }
 }