+++ crates/client/rpc/src/methods/read/get_transaction_by_hash.rs
@@ -48,14 +48,12 @@ pub fn get_transaction_by_hash(starknet: &Starknet, transaction_hash: Felt) -> S
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::test_utils::{make_sample_chain_1, open_testing, SampleChain1};
+    use crate::test_utils::{sample_chain_for_block_getters, SampleChainForBlockGetters};
     use rstest::rstest;
 
     #[rstest]
-    fn test_get_transaction_by_hash() {
-        let _ = env_logger::builder().is_test(true).try_init();
-        let (backend, rpc) = open_testing();
-        let SampleChain1 { tx_hashes, expected_txs, .. } = make_sample_chain_1(&backend);
+    fn test_get_transaction_by_hash(sample_chain_for_block_getters: (SampleChainForBlockGetters, Starknet)) {
+        let (SampleChainForBlockGetters { tx_hashes, expected_txs, .. }, rpc) = sample_chain_for_block_getters;
 
         // Block 0
         assert_eq!(get_transaction_by_hash(&rpc, tx_hashes[0]).unwrap(), expected_txs[0]);
@@ -68,5 +66,13 @@ mod tests {
 
         // Pending
         assert_eq!(get_transaction_by_hash(&rpc, tx_hashes[3]).unwrap(), expected_txs[3]);
+    }
+
+    #[rstest]
+    fn test_get_transaction_by_hash_not_found(sample_chain_for_block_getters: (SampleChainForBlockGetters, Starknet)) {
+        let (SampleChainForBlockGetters { .. }, rpc) = sample_chain_for_block_getters;
+
+        let does_not_exist = Felt::from_hex_unchecked("0x7128638126378");
+        assert_eq!(get_transaction_by_hash(&rpc, does_not_exist), Err(StarknetRpcApiError::TxnHashNotFound));
     }
 }