+++ crates/client/rpc/src/methods/read/get_transaction_status.rs
@@ -55,14 +55,12 @@ pub fn get_transaction_status(starknet: &Starknet, transaction_hash: Felt) -> St
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::test_utils::{make_sample_chain_1, open_testing, SampleChain1};
+    use crate::test_utils::{sample_chain_for_block_getters, SampleChainForBlockGetters};
     use rstest::rstest;
 
     #[rstest]
-    fn test_get_transaction_status() {
-        let _ = env_logger::builder().is_test(true).try_init();
-        let (backend, rpc) = open_testing();
-        let SampleChain1 { tx_hashes, .. } = make_sample_chain_1(&backend);
+    fn test_get_transaction_status(sample_chain_for_block_getters: (SampleChainForBlockGetters, Starknet)) {
+        let (SampleChainForBlockGetters { tx_hashes, .. }, rpc) = sample_chain_for_block_getters;
 
         // Block 0
         assert_eq!(
@@ -87,5 +85,13 @@ mod tests {
             get_transaction_status(&rpc, tx_hashes[3]).unwrap(),
             TransactionStatus::AcceptedOnL2(TransactionExecutionStatus::Succeeded)
         );
+    }
+
+    #[rstest]
+    fn test_get_transaction_status_not_found(sample_chain_for_block_getters: (SampleChainForBlockGetters, Starknet)) {
+        let (SampleChainForBlockGetters { .. }, rpc) = sample_chain_for_block_getters;
+
+        let does_not_exist = Felt::from_hex_unchecked("0x7128638126378");
+        assert_eq!(get_transaction_status(&rpc, does_not_exist), Err(StarknetRpcApiError::TxnHashNotFound));
     }
 }